test: consolidate timerService tests
